Oppdatering i transfer objects og annet
