+++ apps/jenkins-batch-status-service/src/main/java/no/nav/registre/testnorge/jenkinsbatchstatusservice/consumer/command/UpdateOrganisasjonBestillingCommand.java
@@ -2,7 +2,7 @@ package no.nav.registre.testnorge.jenkinsbatchstatusservice.consumer.command;
 
 import lombok.RequiredArgsConstructor;
 import no.nav.testnav.libs.commands.utils.WebClientFilter;
-import no.nav.testnav.libs.dto.organiasjonbestilling.v1.OrderDTO;
+import no.nav.testnav.libs.dto.organisajonbestilling.v1.OrderDTO;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.client.WebClient;
