+++ apps/organisasjon-bestilling-service/src/main/java/no/nav/testnav/apps/organisasjonbestillingservice/controller/v2/OrderControllerV2.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import no.nav.testnav.apps.organisasjonbestillingservice.domain.v2.Order;
 import no.nav.testnav.apps.organisasjonbestillingservice.service.v2.OrderServiceV2;
-import no.nav.testnav.libs.dto.organiasjonbestilling.v2.OrderDTO;
-import no.nav.testnav.libs.dto.organiasjonbestilling.v2.StatusDTO;
+import no.nav.testnav.libs.dto.organisajonbestilling.v2.OrderDTO;
+import no.nav.testnav.libs.dto.organisajonbestilling.v2.StatusDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
