+++ apps/organisasjon-mottak-service/src/main/java/no/nav/registre/testnorge/organisasjonmottak/consumer/command/RegisterBestillingCommand.java
@@ -2,7 +2,7 @@ package no.nav.registre.testnorge.organisasjonmottak.consumer.command;
 
 import lombok.RequiredArgsConstructor;
 import no.nav.testnav.libs.commands.utils.WebClientFilter;
-import no.nav.testnav.libs.dto.organiasjonbestilling.v2.OrderDTO;
+import no.nav.testnav.libs.dto.organisajonbestilling.v2.OrderDTO;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.client.WebClient;
