--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/applikasjonsanalyseservice/v1/ApplicationInfoDTO.java
@@ -1,19 +1,0 @@
-package no.nav.testnav.libs.dto.applikasjonsanalyseservice.v1;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-import java.util.Set;
-
-@Value
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class ApplicationInfoDTO {
-    String name;
-    String cluster;
-    String namespace;
-    Set<DependencyDTO> dependencies;
-}
