--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/dependencyanalysis/v1/DependencyDTO.java
@@ -1,19 +1,0 @@
-package no.nav.testnav.libs.dto.dependencyanalysis.v1;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-@Value
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class DependencyDTO {
-
-    @JsonProperty(required = true)
-    String name;
-    @JsonProperty(required = true)
-    Boolean external;
-}
