--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/hendelse/v1/HendelseType.java
@@ -1,6 +1,0 @@
-package no.nav.testnav.libs.dto.hendelse.v1;
-
-public enum HendelseType {
-    SYKEMELDING_OPPRETTET,
-    ARBEIDSFORHOLD_OPPRETTET
-}
