--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/organiasjonbestilling/v1/OrderDTO.java
@@ -1,15 +1,0 @@
-package no.nav.testnav.libs.dto.organiasjonbestilling.v1;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-@Value
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class OrderDTO {
-    String miljo;
-    Long batchId;
-}
