+++ libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/organisajonbestilling/v1/Status.java
@@ -1,4 +1,4 @@
-package no.nav.testnav.libs.dto.organiasjonbestilling.v1;
+package no.nav.testnav.libs.dto.organisajonbestilling.v1;
 
 public enum Status {
     NOT_STARTED,
