--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/rapprtering/v1/EntryStatus.java
@@ -1,7 +1,0 @@
-package no.nav.testnav.libs.dto.rapprtering.v1;
-
-public enum EntryStatus {
-    INFO,
-    WARNING,
-    ERROR
-}
