--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/samhandlerregisteret/v1/IdentDTO.java
@@ -1,18 +1,0 @@
-package no.nav.testnav.libs.dto.samhandlerregisteret.v1;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-@Value
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class IdentDTO {
-    @JsonProperty("ident")
-    private final String ident;
-    @JsonProperty("ident_type_kode")
-    private final String identTypeKode;
-}
