--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/statistikkservice/v1/StatistikkType.java
@@ -1,9 +1,0 @@
-package no.nav.testnav.libs.dto.statistikkservice.v1;
-
-public enum StatistikkType {
-    ANTALL_ARBEIDSTAKERE_SYKEMELDT,
-    ANTALL_ARBEIDSTAKERE_SOM_ER_I_ARBEIDSSTYRKEN,
-    ANDEL_PERSONER_ALDER_0_19,
-    ANDEL_PERSONER_ALDER_20_64,
-    ANDEL_PERSONER_ALDER_OVER_65
-}
