--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/statistikkservice/v1/StatistikkValueType.java
@@ -1,6 +1,0 @@
-package no.nav.testnav.libs.dto.statistikkservice.v1;
-
-public enum StatistikkValueType {
-    PERCENTAGE,
-    COUNT
-}
