--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/v1/PersonDTO.java
@@ -1,23 +1,0 @@
-package no.nav.testnav.libs.dto.v1;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.Value;
-
-import java.util.Set;
-
-@Value
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class PersonDTO {
-    String ident;
-    String fornavn;
-    String etternavn;
-    String adresse;
-    String postnr;
-    String by;
-    String opprinnelse;
-    Set<String> tags;
-}
