--- libs/data-transfer-objects/src/main/java/no/nav/testnav/libs/dto/v1/Persondatasystem.java
@@ -1,6 +1,0 @@
-package no.nav.testnav.libs.dto.v1;
-
-public enum Persondatasystem {
-    TPS,
-    PDL
-}
